% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{mvbsc}
\alias{mvbsc}
\title{...}
\usage{
mvbsc(codes, wt, cosK, h, k, R, Z0, seed = 123, ...)
}
\arguments{
\item{codes}{...}

\item{wt}{...}

\item{cosK}{...}

\item{h}{...}

\item{k}{...}

\item{R}{...}

\item{Z0}{...}

\item{seed}{...}

\item{...}{...}
}
\description{
...
}
