% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{semi_kmeans}
\alias{semi_kmeans}
\title{...}
\usage{
semi_kmeans(W, k, constraint_set, alpha, beta, seed)
}
\arguments{
\item{W}{UU'}

\item{k}{...}

\item{constraint_set}{...}

\item{alpha}{...}

\item{beta}{...}

\item{seed}{...}
}
\description{
...
}
