% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{mvbsc}
\alias{mvbsc}
\title{...}
\usage{
mvbsc(codes, distance, similarity, k, delta = NULL, h = NULL,
  wt = NULL, seed = 123)
}
\arguments{
\item{codes}{...}

\item{distance}{...}

\item{similarity}{...}

\item{k}{...}

\item{delta}{...}

\item{h}{...}

\item{wt}{...}

\item{seed}{...}
}
\description{
...
}
