% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{mvbsc_fit}
\alias{mvbsc_fit}
\title{...}
\usage{
mvbsc_fit(codes, distance, similarity, ncluster, weights = NULL,
  delta = NULL, band = NULL, seed = 123)
}
\arguments{
\item{codes}{...}

\item{distance}{...}

\item{similarity}{...}

\item{ncluster}{...}

\item{weights}{...}

\item{delta}{...}

\item{band}{...}

\item{seed}{...}
}
\description{
...
}
