% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{mvbsc_fit}
\alias{mvbsc_fit}
\title{...}
\usage{
mvbsc_fit(codes, distance, similarity, k, delta, h, wt, seed)
}
\arguments{
\item{codes}{...}

\item{distance}{...}

\item{similarity}{...}

\item{k}{...}

\item{delta}{...}

\item{h}{...}

\item{wt}{...}

\item{seed}{...}
}
\description{
...
}
