% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBSC.R
\name{mvbsc_fit0}
\alias{mvbsc_fit0}
\title{...}
\usage{
mvbsc_fit0(codes, distance, similarity, ncluster, weights, delta, band,
  seed)
}
\arguments{
\item{codes}{...}

\item{distance}{...}

\item{similarity}{...}

\item{ncluster}{...}

\item{weights}{...}

\item{delta}{...}

\item{band}{...}

\item{seed}{...}
}
\description{
...
}
